+++ src/database/projects.js
@@ -1,0 +1,43 @@
+const projects = [
+  {
+    id: 1,
+    project_name: 'project One',
+    description:
+      'Lorem ipsum dolor sit amet, consectetur elit. Vivamus imperdiet ut quam sit amet vehicula.',
+    cover_image:
+      'https://images.unsplash.com/photo-1572177812156-58036aae439c?ixlib=rb-1.2.1&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1470&q=80',
+    repo_link: '#',
+    keywords: ['keyword1', 'keyword2'],
+    tags: ['lang1', 'lang2'],
+    tech_stack: ['tech1', 'tech2'],
+    demo: '#',
+  },
+  {
+    id: 2,
+    project_name: 'project Two',
+    description:
+      'Lorem ipsum dolor sit amet, consectetur elit. Vivamus imperdiet ut quam sit amet vehicula.',
+    cover_image:
+      'https://images.unsplash.com/photo-1542621334-a254cf47733d?ixlib=rb-1.2.1&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1470&q=80',
+    repo_link: '#',
+    keywords: ['keyword1', 'keyword2'],
+    tags: ['lang1', 'lang2'],
+    tech_stack: ['tech1', 'tech2'],
+    demo: '#',
+  },
+  {
+    id: 4,
+    project_name: 'project four',
+    description:
+      'Lorem ipsum dolor sit amet, consectetur elit. Vivamus imperdiet ut quam sit amet vehicula.',
+    cover_image:
+      'https://images.unsplash.com/photo-1630305107616-64fde2a71808?ixlib=rb-1.2.1&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=1470&q=80',
+    repo_link: '#',
+    keywords: ['keyword1', 'keyword2'],
+    tags: ['lang1', 'lang2'],
+    tech_stack: ['tech1', 'tech2'],
+    demo: '#',
+  },
+];
+
+export default projects;
